requirements.txt
strawberry-graphql[opentelemetry,pydantic]==0.73.9
databases[sqlite]==0.4.3
# sqlalchemy version set by databases
sqlalchemy
# starlette version set by strawberry
starlette
uvicorn[standard]==0.13.4
